feat: extract AiModelSelect component, remove model field duplication
Single AiModelSelect component now owns the model label, spinner,
Select/Input toggle, and hint text. Consumed by ai-settings-dialog,
settings/page, and setup-wizard — removes ~60 lines of duplicated
fetch + render logic.
